Move common notification code to a standalone file
