+++ carbon/src/commonMain/kotlin/com/gabrieldrn/carbon/notification/CalloutNotification.kt
@@ -16,30 +16,18 @@
 
 package com.gabrieldrn.carbon.notification
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.BasicText
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawWithContent
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.unit.dp
 import com.gabrieldrn.carbon.Carbon
 import com.gabrieldrn.carbon.foundation.spacing.SpacingScale
-import com.gabrieldrn.carbon.icons.CheckmarkFilledIcon
-import com.gabrieldrn.carbon.icons.ErrorFilledIcon
-import com.gabrieldrn.carbon.icons.InformationFilledIcon
-import com.gabrieldrn.carbon.icons.WarningFilledIcon
 
 /**
  * # Callout notification
@@ -74,14 +62,10 @@ public fun CalloutNotification(
     )
 
     NotificationContainer(
+        status = status,
         colors = colors,
         modifier = modifier
     ) {
-        Icon(
-            status = status,
-            colors = colors,
-        )
-
         FlowRow(
             horizontalArrangement = Arrangement.spacedBy(4.dp),
             verticalArrangement = Arrangement.spacedBy(2.dp),
@@ -138,72 +122,5 @@ public fun CalloutNotification(
         modifier = modifier,
         title = title,
         highContrast = highContrast
-    )
-}
-
-private val iconSize = 20.dp
-
-@Composable
-private fun Icon(
-    status: NotificationStatus,
-    colors: NotificationColors,
-    modifier: Modifier = Modifier
-) {
-    when (status) {
-        NotificationStatus.Informational -> InformationFilledIcon(
-            tint = colors.iconColor,
-            innerTint = colors.iconInnerColor,
-            size = iconSize,
-            modifier = modifier.testTag(NotificationTestTags.ICON_INFORMATIONAL)
-        )
-        NotificationStatus.Success -> CheckmarkFilledIcon(
-            tint = colors.iconColor,
-            size = iconSize,
-            modifier = modifier.testTag(NotificationTestTags.ICON_SUCCESS)
-        )
-        NotificationStatus.Warning -> WarningFilledIcon(
-            tint = colors.iconColor,
-            innerTint = colors.iconInnerColor,
-            size = iconSize,
-            modifier = modifier.testTag(NotificationTestTags.ICON_WARNING)
-        )
-        NotificationStatus.Error -> ErrorFilledIcon(
-            tint = colors.iconColor,
-            size = iconSize,
-            modifier = modifier.testTag(NotificationTestTags.ICON_ERROR)
-        )
-    }
-}
-
-@Composable
-private fun NotificationContainer(
-    colors: NotificationColors,
-    modifier: Modifier = Modifier,
-    content: @Composable RowScope.() -> Unit = {}
-) {
-    Row(
-        modifier = modifier
-            .background(colors.backgroundColor)
-            .drawWithContent {
-                drawContent()
-                val contourWidthPx = 1.dp.toPx()
-                drawRect(
-                    brush = SolidColor(colors.contourColor),
-                    topLeft = Offset(contourWidthPx / 2, contourWidthPx / 2),
-                    size = Size(
-                        size.width - contourWidthPx,
-                        size.height - contourWidthPx
-                    ),
-                    style = Stroke(1.dp.toPx())
-                )
-                drawRect(
-                    brush = SolidColor(colors.borderLeftColor),
-                    topLeft = Offset.Zero,
-                    size = Size(3f.dp.toPx(), size.height)
-                )
-            }
-            .padding(SpacingScale.spacing05)
-            .testTag(NotificationTestTags.CONTAINER),
-        content = content
     )
 }
